requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
openai==0.6.3
pkg-resources==0.0.0
python-dotenv==0.17.0
requests==2.25.1
tqdm==4.60.0
urllib3==1.26.4
